query: reorganize files
